requirements.txt
hikari-lightbulb==0.0.59